+++ src/main/resources/application.properties
@@ -16,7 +16,6 @@ cryostat.discovery.plugins.jwt.signature.algorithm=HS256
 cryostat.discovery.plugins.jwt.encryption.algorithm=dir
 cryostat.discovery.plugins.jwt.encryption.method=A256GCM
 cryostat.discovery.kubernetes.enabled=false
-cryostat.discovery.kubernetes2.enabled=false
 cryostat.discovery.kubernetes.port-names=
 cryostat.discovery.kubernetes.port-numbers=
 cryostat.discovery.kubernetes.namespaces=
